Добавлена остановка дальнейшего разбора в случае семантической ошибки

+++ lex.yy.cpp
@@ -2291,32 +2291,34 @@ int main(int argc, char* argv[])
             yyin = stdin;
 
     printf("Start parsing\n");
-    yyparse();
-    printf("Successful parsing\n");
+    int parseRes = yyparse();
+    if (parseRes == 0) {
+        printf("Successful parsing\n");
 
-    printf("Start generating dot file before semantic analysis\n");
-    string res = root->toDot();
-    ofstream out("dot_before.dot");
-    out << res;
-    out.close();
-    printf("Successful generating dot file before semantic analysis\n");
-    
-    printf("Start semantic analysis\n");
-    root->fillClassesTable();
-    printf("Successful semantic analysis\n");
+        printf("Start generating dot file before semantic analysis\n");
+        string res = root->toDot();
+        ofstream out("dot_before.dot");
+        out << res;
+        out.close();
+        printf("Successful generating dot file before semantic analysis\n");
+        
+        printf("Start semantic analysis\n");
+        root->fillClassesTable();
+        printf("Successful semantic analysis\n");
 
-    printf("Start printing tables after semantic analysis\n");
-    filesystem::path p = filesystem::path("./out/");
-    filesystem::remove_all(p);
-    filesystem::create_directory(p);
-    ClassesTable::toCsvFile("./out/");
-    FunctionsTable::toCsvFile("FunctionsTable.csv", "./out/");
-    printf("Successful printing tables after semantic analysis\n");
+        printf("Start printing tables after semantic analysis\n");
+        filesystem::path p = filesystem::path("./out/");
+        filesystem::remove_all(p);
+        filesystem::create_directory(p);
+        ClassesTable::toCsvFile("./out/");
+        FunctionsTable::toCsvFile("FunctionsTable.csv", "./out/");
+        printf("Successful printing tables after semantic analysis\n");
 
-    printf("Start generating dot file after semantic analysis\n");
-    res = root->toDot();
-    out.open("dot_after.dot");
-    out << res;
-    out.close();
-    printf("Successful generating dot file after semantic analysis\n");
+        printf("Start generating dot file after semantic analysis\n");
+        res = root->toDot();
+        out.open("dot_after.dot");
+        out << res;
+        out.close();
+        printf("Successful generating dot file after semantic analysis\n");
+    }
 }
